move mkldnn test functions to one file and use DNN_LOG_LEVEL
